lang-portal/backend-flask/requirements.txt
flask==3.0.0
flask-cors==4.0.0
openai==1.71.0
python-dotenv==1.0.1
requests>=2.32.3
pytest==8.0.0
pytest-flask==1.3.0
google-search-results
aiohttp==3.9.3
beautifulsoup4==4.12.3
instructor==1.7.9
httpx>=0.24.1 